% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_chr_to_factor.R
\name{all_chr_to_factor}
\alias{all_chr_to_factor}
\title{Convert all character variables to factors}
\usage{
all_chr_to_factor(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Convert all character variables to factors
}
\examples{
library(BioStat)

# Basic syntax
all_chr_to_factor(PlantGrowth)

# Update the same object using operator from `magrittr` package
PlantGrowth \%<>\% all_chr_to_factor()


# Let's create a data frame
df <- data.frame(letters  = letters[1:5],
                 letters2 = LETTERS[1:5],
                 stringsAsFactors = FALSE)

mapply(class, df) # show classes of columns

# Let's convert all character variables to strings
df2 <- all_chr_to_factor(df)

# Let's check the classes in each column
mapply(class, df2)

# Works with tibbles too
tbl  <- as.tibble(df)
tbl2 <- all_chr_to_factor(tbl)
mapply(class, tbl2)
}
