% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_BioStat-package.R
\docType{package}
\name{BioStat}
\alias{BioStat}
\alias{BioStat-package}
\title{A Collection of Functions for Biostatistics and Biometry Lectures

The package contains convenience functions for common statistical analyses
and data wrangling procedures.}
