% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc__cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\alias{cohens_d.default}
\title{Cohen's Effect Size}
\usage{
cohens_d(x, y = NULL, pooled = TRUE, correct = FALSE, conf.level = NA,
  na.rm = FALSE, group = NULL, ...)

\method{cohens_d}{default}(x, y = NULL, pooled = TRUE, correct = FALSE,
  conf.level = NA, na.rm = FALSE, group = NULL, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{pooled}{(logical) Flag indicating whether compute pooled standard deviation
or the whole sample standard deviation. Default is \code{TRUE}.}

\item{correct}{(logical) Flag indicating whether to apply the Hedges correction.
(Default: \code{FALSE})}

\item{conf.level}{(numeric|\code{NA}) confidence level of the interval. Set this to \code{NA}, if no
confidence intervals should be calculated. (This is the default)}

\item{na.rm}{(logical) Should missing values be removed? Defaults to \code{FALSE}.}
}
\value{
a numeric vector with 3 elements: \item{d}{the effect size d}
\item{lwr.ci}{lower bound of the confidence interval} \item{upr.ci}{upper
bound of the confidence interval}
}
\description{
Computes the Cohen's d and Hedges'g effect size statistics.
}
\section{Origin of the code}{


The main part of code and documentation of this function is imported from: \cr
    Andri Signorell et al. (2017).
    \pkg{DescTools}: Tools for descriptive statistics.
    R package version \emph{0.99.22} (2017-09-14).

The modifications were made by Vilmantas Gegzna, but the logic of statistical
computations were not changed.
}

\examples{
data("d.pizza", package = "DescTools")
library(BioStat)

x <- d.pizza$price[d.pizza$driver == "Carter"]
y <- d.pizza$price[d.pizza$driver == "Miller"]

cohens_d(x, y, conf.level = 0.95, na.rm = TRUE)


}
\references{
Cohen, J. (1988) \emph{Statistical power analysis for the
behavioral sciences (2nd ed.)} Academic Press, New York.

Hedges, L. V. & Olkin, I. (1985) \emph{Statistical methods for
meta-analysis} Academic Press, Orlando, FL

Smithson, M.J. (2003) \emph{Confidence Intervals, Quantitative Applications
in the Social Sciences Series}, No. 140. Thousand Oaks, CA: Sage. pp. 39-41
}
\seealso{
\code{\link{mean}}, \code{\link{var}}
}
\author{
Andri Signorell <andri@signorell.net>
}
\keyword{effect}
\keyword{size}
