% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmisc_describe.R
\name{describe_data}
\alias{describe_data}
\title{describe}
\usage{
describe_data(x, ...)
}
\arguments{
\item{x}{object to make a statistical summary for}

\item{...}{other parameters}
}
\description{
From "Hmisc"
}
