% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers_in_columns.R
\name{format_numbers_in_columns}
\alias{format_numbers_in_columns}
\title{Format numbers in columns}
\usage{
format_numbers_in_columns(data, digits = 3, format = "f", fun = formatC,
  ...)
}
\arguments{
\item{data}{A data-frame-like data.}

\item{digits}{(Either one integer or a vector of inegers for each column) a desired number of digits after the decimal point (if \code{format} = "f") or a number of significant digits (\code{format} = "g", = "e" or = "fg"). \cr}

\item{format}{either one value or a vector of values for each column.}

\item{fun}{A function that does the formatting. Default is \code{\link[base]{formatC}}.}

\item{...}{Other parameters to be passed to \code{fun}.}
}
\description{
Function formats only columns. Other classes are left intact.
}
\examples{
library(BioStat)
DATA <- head(iris)

# The same rounding for each column
format_numbers_in_columns(DATA, 2)

# Different roundingfor each column
format_numbers_in_columns(DATA, c(2,2,3,3,NA))

# Use NA to skip formatiing of the column
format_numbers_in_columns(DATA, c(4,NA,3,NA,NA))
}
