% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head_tail}
\alias{head_tail}
\title{[!] Show several first and last rows of a data frame}
\usage{
head_tail(obj, n = 4, top = n, bottom = n, sep = "...")
}
\arguments{
\item{obj}{A data frame.}

\item{n}{(integer) Number of top and bottom rows to display.}

\item{top}{(integer) Number of top rows to display.}

\item{bottom}{(integer) Number of bottom rows to display.}

\item{sep}{(character) Separator between displayed top and bottom lines.}
}
\value{
A truncated data frame (which is intended to be printed) with all
        variables converted to strings.
}
\description{
[!] Show several first and last rows of a data frame
}
\examples{
data(swiss)
head_tail(swiss)
}
\keyword{utilities}
