% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cld.R
\name{make_cld}
\alias{make_cld}
\alias{make_cld.posthocTGH}
\title{Make a compact letter display}
\usage{
make_cld(obj, ...)

\method{make_cld}{posthocTGH}(obj, ...)
}
\arguments{
\item{obj}{Object with post-hoc results}

\item{...}{Further arguments to methods.}
}
\value{
(A dataframe with) compact letter display.
}
\description{
Make a compact letter display for results of ANOVA post-hoc test.
}
