% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{pearson.test}
\alias{pearson.test}
\title{Pearson chi-square test for normality with formula interface}
\usage{
pearson.test(x, n.classes = ceiling(2 * (n^(2/5))), adjust = TRUE,
  data = NULL, groups = NULL)
}
\arguments{
\item{x}{a numeric vector of data values. Missing values are allowed.}

\item{n.classes}{The number of classes. The default is due to Moore (1986).}

\item{adjust}{logical; if \code{TRUE} (default), the p-value is computed from
  a chi-square distribution with \code{n.classes}-3 degrees of freedom, otherwise
  from a chi-square distribution with \code{n.classes}-1 degrees of freedom.}

\item{data}{a data frame.
Note that the default is \code{data=parent.frame()}.  This makes it convenient to
use this function interactively by treating the working envionment as if it were 
a data frame.  But this may not be appropriate for programming uses.  
When programming, it is best to use an explicit \code{data} argument
-- ideally supplying a data frame that contains the variables mentioned
in \code{formula}.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}
}
\value{
...
}
\description{
Performs the Pearson chi-square test for the composite
hypothesis of normality, see e.g. Thode (2002, Sec. 5.2).
More information in: \code{\link[nortest]{pearson.test}}.
}
\examples{
library(BioStat)
data(CO2)
pearson.test(uptake ~ Type + Treatment, data = CO2)

}
