% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocTGH.R
\name{posthocTGH}
\alias{posthocTGH}
\alias{print.posthocTGH}
\title{Games-Howell and Tukey post-hoc tests}
\usage{
posthocTGH(y, x, method = c("games-howell", "tukey"), conf.level = 0.95,
  digits = 2, p.adjust = "holm", formatPvalue = TRUE)

\method{print}{posthocTGH}(x, digits = x$input$digits, ...)
}
\arguments{
\item{y}{y (numeric) a numeric vector.}

\item{x}{x (factor) a vector that either is a factor or can be converted into
one.}

\item{method}{(\code{"games-howell"}|\code{"tukey"}) Which post-hoc
tests to conduct. Valid values are "tukey" and "games-howell".}

\item{conf.level}{Confidence level of the confidence intervals.}

\item{digits}{The number of digits to show in the output.}

\item{p.adjust}{Any valid \code{\link{p.adjust}} method.}

\item{formatPvalue}{Whether to format the p values according to APA
standards (i.e. replace all values lower than .001 with '<.001'). This only
applies to the printing of the object, not to the way the p values are
stored in the object.}
}
\value{
A list of three elements:
\item{input}{List with input arguments}
\item{intermediate}{List of intermediate objects.}
\item{output}{List with
two objects 'tukey' and 'games.howell', containing the outcomes for the
respective post-hoc tests.}
}
\description{
Either Games-Howell or Tukey post-hoc tests for one-way analysis of
variance (ANOVA).
}
\details{
This function is omported from package \pkg{userfriendlyscience}
(\code{\link[userfriendlyscience]{posthocTGH}})
}
\note{
This function is based on a file that was once hosted at
http://www.psych.yorku.ca/cribbie/6130/games_howell.R, but has been removed
since. It was then adjusted for implementation in the
\code{\link{userfriendlyscience}} package. Jeffrey Baggett needed the
confidence intervals, and so emailed them, after which his updated functions
was used. In the meantime, it appears Aaron Schlegel
(\url{https://rpubs.com/aaronsc32}) independently developed a version with
confidence intervals and posted it on RPubs at
\url{https://rpubs.com/aaronsc32/games-howell-test}.
}
\examples{
library(BioStat)

### Compute post-hoc statistics using the Games-Howell method
posthocTGH(y=ChickWeight$weight, x=ChickWeight$Diet)

### Compute post-hoc statistics using the Tukey method
posthocTGH(y=ChickWeight$weight, x=ChickWeight$Diet, method="tukey");

}
\seealso{
\itemize{
\item{\url{http://www.gcf.dkf.unibe.ch/BCB/files/BCB_10Jan12_Alexander.pdf}}
\item{\url{https://rpubs.com/aaronsc32/games-howell-test}}
\item{\url{http://aoki2.si.gunma-u.ac.jp/R/src/tukey.R}}
\item{\url{https://stats.stackexchange.com/questions/83941/games-howell-post-hoc-test-in-r}}

}
}
\author{
Gjalt-Jorn Peters (Open University of the Netherlands) & Jeff Bagget
(University of Wisconsin - La Crosse)

Maintainer: Gjalt-Jorn Peters <gjalt-jorn@userfriendlyscience.com>
}
\keyword{utilities}
