% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_p.R
\name{prettify_p_value}
\alias{prettify_p_value}
\alias{prettify_p_column}
\title{Format p-values}
\usage{
prettify_p_value(x)

prettify_p_column(data, colname = "p.value", prettify = TRUE, ...)
}
\arguments{
\item{x}{A p-value (numeric or coercible to numeric).}

\item{data}{A data frame.}

\item{colname}{(string) A name of column with p-values.}

\item{prettify}{(logical) should function \code{prettify_p_value}
be applied.}

\item{...}{(not used yet)}
}
\description{
Values below 0.001 are printed as \code{"<0.001"},
between [0.001 and 0.01) as \code{"<0.01"},
all others are rounded to two decimal places. \cr
}
\details{
\code{prettify_p_column} formats a column of p-values in a data frame.\cr

\code{prettify_p_column} formats a single value.\cr
}
\examples{
library(BioStat)

# Make a dataframe with p-values
rez <- test_normality(uptake ~ Type,
                      data = CO2)
rez

# Shapiro-Wilk normality test
#
#          Type  statistic  p.value
# 1      Quebec     0.860   0.000111
# 2 Mississippi     0.936   0.0213

# Do prettification:
prettify_p_column(rez)

# Shapiro-Wilk normality test
#
#           Type statistic p.value
# 1      Quebec     0.860  <0.001
# 2 Mississippi     0.936   0.02

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Prettify single p-values

prettify_p_value(0.005)
prettify_p_value(0.0005)
prettify_p_value(0.052147)
prettify_p_value("0.0022")
prettify_p_value("0.052")

}
