% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers_in_columns.R
\name{pretty_p}
\alias{pretty_p}
\title{Pretty formatting of a p-value}
\usage{
pretty_p(x)
}
\arguments{
\item{x}{A p value.}
}
\value{
A string with p value in a pretty format.
}
\description{
Pretty formatting of a p-value
}
\examples{
pretty_p(0.005)

pretty_p(0.0005)

pretty_p(x = 0.052147)


}
