% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers_in_columns.R
\name{pretty_p_values}
\alias{pretty_p_values}
\title{Format column of p-values in a data frame}
\usage{
pretty_p_values(data, name = "p.value", pretty_p = TRUE, ...)
}
\arguments{
\item{data}{A data frame}

\item{name}{A name of column with p-values}

\item{pretty_p}{(logical) should function \code{pretty_p} be applied.}

\item{...}{(not used yet)}
}
\description{
Format column of p-values in a data frame
}
\examples{
rez <- test_normality(uptake ~ Type + Treatment, data = CO2)

rez

pretty_p_values(rez)

}
