% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{print.ntest}
\alias{print.ntest}
\title{Print `ntest` object}
\usage{
\method{print}{ntest}(x, ...)
}
\arguments{
\item{x}{- object of class \code{ntest}.}

\item{...}{}
}
\description{
Print `ntest` object
}
