% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_data.R
\name{qq_data}
\alias{qq_data}
\alias{coef.qqdata}
\title{Calculate data for a qq-plot}
\usage{
qq_data(x, distribution = "norm", ..., envelope = 0.95,
  line = c("quartiles", "robust", "int=0,slope=1"), labels = NULL,
  groups = NULL, data = NULL, method = if (distribution == "norm")
  "normal" else "any")

\method{coef}{qqdata}(obj)
}
\arguments{
\item{x}{Either a formula or a numeric vector or a name of a vector
in \code{data}.\cr
If \code{x} is a formula (e.g. \code{variable ~ factor}), left
side provides variable to be summarized. Right side and condition
describe subsets. If the left side is empty, right side and
condition are shifted over as a convenience.}

\item{distribution}{root name of comparison distribution -- e.g., \code{"norm"} for the
    normal distribution; \code{t} for the t-distribution.}

\item{...}{Parameters to be passed to function, selected in \code{distribution}}

\item{envelope}{confidence level for point-wise confidence envelope, or 
    \code{FALSE} for no envelope.}

\item{line}{\code{"quartiles"} to pass a line through the quartile-pairs, or
    \code{"robust"} for a robust-regression line; the latter uses the \code{rlm}
    function in the \code{MASS} package. Specifying \code{line = "none"} suppresses the line.}

\item{labels}{vector of text strings to be used to identify points, defaults to
    \code{names(x)} or observation numbers if \code{names(x)} is \code{NULL}.}

\item{data}{A data frame that contains the variables mentioned in \code{x}.}

\item{obj}{qqdata object.}
}
\value{
An object, which inherits from classes \code{qqdata} and
        \code{data.frame}. The object contains information, needed
        to plot a qqplot with reference line and its confidence intervals.
        These variables are contained: \itemize{
        \item\strong{x} – x axis values;
        \item\strong{y} – y axis values for points of qq plot;
        \item\strong{labels} – labels for each point;
        \item\strong{ref_y} – y axis values for reference line
        \item\strong{ref_ci_upper} and \strong{ref_ci_lower}
          – y axis values for upper and lower pointwise
           confidence interval of a reference line.
        }
}
\description{
Calculate data for a qq-plot
}
\examples{

library(BioStat)
data(chickwts, package = "datasets")

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as formula + data:

QQdata <- qq_data(~weight, data = chickwts)
head(QQdata)
coef(QQdata)


# Column ".group" is added if applied by group:

QQdata <- qq_data(weight ~ feed, data = chickwts)
head(QQdata)
coef(QQdata)

qq_plot(weight ~ feed, data = chickwts)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as variable name + data:

QQdata <- qq_data("weight", data = chickwts)
head(QQdata)
coef(QQdata)

QQdata <- qq_data("weight",groups = "feed", data = chickwts)
head(QQdata)
coef(QQdata)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as vector

QQdata <- qq_data(chickwts$weight)
head(QQdata)
coef(QQdata)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as vector, several groups.
# Column ".group" is added

QQdata <- qq_data(chickwts$weight, groups = chickwts$feed)
head(QQdata)
coef(QQdata)



}
\seealso{
\code{\link[car]{qqPlot}} in \pkg{car} package,
 \code{\link[stats]{qqplot}} in \pkg{stats} package.
}
