% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_p.R
\name{rm_zero}
\alias{rm_zero}
\title{Function to remove zero at the beginning of a number}
\usage{
rm_zero(str)
}
\arguments{
\item{str}{a string or convertible to string}
}
\value{
A value without leading zero
}
\description{
Function to remove zero at the beginning of a number
}
\examples{

rm_zero(0.02)

}
\keyword{internal}
