% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{sf.test}
\alias{sf.test}
\title{Shapiro-Francia test for normality with formula interface}
\usage{
sf.test(x, ..., data = NULL, groups = NULL)
}
\arguments{
\item{x}{a numeric vector of data values, the number of 
  which must be between 5 and 5000. Missing values are allowed.}

\item{...}{additional arguments passed to \code{FUN}}

\item{data}{a data frame.
Note that the default is \code{data=parent.frame()}.  This makes it convenient to
use this function interactively by treating the working envionment as if it were 
a data frame.  But this may not be appropriate for programming uses.  
When programming, it is best to use an explicit \code{data} argument
-- ideally supplying a data frame that contains the variables mentioned
in \code{formula}.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}
}
\value{
...
}
\description{
Performs the Shapiro-Francia test for the composite
hypothesis of normality, see e.g. Thode (2002, Sec. 2.3.2).
More information in: \code{\link[nortest]{sf.test}}.
}
\examples{
library(BioStat)
data(CO2)
sf.test(uptake ~ Type + Treatment, data = CO2)

}

