% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{shapiro.test}
\alias{shapiro.test}
\title{Normality tests with formula interface}
\usage{
shapiro.test(x, ..., data = NULL, groups = NULL)
}
\arguments{
\item{x}{a numeric vector of data values. Missing values are allowed,
    but the number of non-missing values must be between 3 and 5000.}

\item{...}{additional arguments passed to \code{FUN}}

\item{data}{a data frame.
Note that the default is \code{data=parent.frame()}.  This makes it convenient to
use this function interactively by treating the working envionment as if it were 
a data frame.  But this may not be appropriate for programming uses.  
When programming, it is best to use an explicit \code{data} argument
-- ideally supplying a data frame that contains the variables mentioned
in \code{formula}.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}
}
\value{
...
}
\description{
Performs the Shapiro-Wilk test of normality.
}
\examples{
library(BioStat)
data(CO2)
shapiro.test(uptake ~ Type + Treatment, data = CO2)

}
