% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_coef.R
\name{standardized_coef}
\alias{standardized_coef}
\alias{print.lm_beta}
\title{Compute Standardized Regression Coefficients}
\usage{
standardized_coef(obj)

\method{print}{lm_beta}(x, ..., digits = 3)
}
\arguments{
\item{obj}{(\code{lm} object) A result of function \code{lm()}.}

\item{x}{\code{lm_beta} object.}

\item{...}{further parameters to \code{print} method.}

\item{digits}{(integer) number of decimal places to round the answer to.
Default is 3.}
}
\value{
Object of class \code{lm_beta} which is a named numeric vector  that represents each standardized coefficient from \code{lm()} model.
}
\description{
Compute the standardized regression coefficients (beta) from an object of class \code{lm}).
}
\details{
This function is inspired by function  \code{\pkg{QuantPsyc}::lm.beta()} written by Thomas D. Fletcher. \cr
\code{standardized_coef()} provides standardized coefficients even when interaction members are present. This is achieved by computing whole model matrix (with all right-hand side members of formula used in call of \code{lm()}) and calculating standard deviations of each regressor (including interaction members) based on these columns. The remaining calculations are the same as in \code{\pkg{QuantPsyc}::lm.beta()}.
}
\examples{

data(USJudgeRatings)
us <- USJudgeRatings
names(us)

lm1 <- lm(CONT ~ INTG + DMNR + log(DILG), data = us)
standardized_coef(lm1)

lm2 <- lm(CONT ~ INTG + DMNR*DILG, data = us)
standardized_coef(lm2)

}
\seealso{
\itemize{
    \item \code{\link[stats]{lm}} in package \pkg{stats}.
    \item \code{\link[QuantPsyc]{lm.beta}} in package \pkg{QuantPsyc}.
}
}
\author{
Vilmantas Gegzna (modified function written by Thomas D. Fletcher).
}
\keyword{models}
