% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_chr_to_factor.R
\name{all_chr_to_factor}
\alias{all_chr_to_factor}
\title{Convert all character variables to factors}
\usage{
all_chr_to_factor(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Convert all character variables to factors
}
\examples{
df <- data.frame(letters  = letters[1:5],
                 letters2 = LETTERS[1:5],
                 stringsAsFactors = FALSE)

mapply(class, df) # show classes of columns

df2 <- all_chr_to_factor(df)
mapply(class, df2)

}
