% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxplot_plus.R
\name{gg_boxplot_plus}
\alias{gg_boxplot_plus}
\title{Plot a boxplot with additional components}
\usage{
gg_boxplot_plus(formula, data = NULL, cld = NULL, sort_groups = c("no",
  "yes", "ascending", "descending"), sort_fun = median, add_points = TRUE,
  add_mean_ci = TRUE, add_median_ci = FALSE, conf_level = 0.95,
  ci_boot_reps = 999, cld_color = "black", cld_y_adj = 0,
  cld_y_mult = 0.05, ci_x_adj = -0.3, points_x_adj = 0.3, ...)
}
\arguments{
\item{formula}{a formula with two variable names to analyze. First one is numeric, second one is a factor, e.g. \code{y ~ group}.}

\item{data}{a data frame with data.}

\item{cld}{a data frame with cld results (object of class \code{cld_object}).}

\item{sort_groups}{(\code{"no"}|\code{"yes"}|\code{"ascending"}|\code{"descending"}) Sort groups by position of median.}

\item{sort_fun}{A function that calculates one numeric statistic
(name without quotes). May be \code{median},
\code{mean}, \code{sd}, \code{var}, \code{IQR}, or similar.}

\item{add_points}{(\code{TRUE}|\code{FALSE})}

\item{add_mean_ci}{(\code{TRUE}|\code{FALSE})}

\item{add_median_ci}{(\code{TRUE}|\code{FALSE})}

\item{conf_level}{(numeric) Confidence level for confidence interval. Number from 0 to 1. Default is 0.95.}

\item{ci_boot_reps}{(numeric) Number of bootstrap repetitions for mean confidence interval calculation.}

\item{cld_color}{(character) Name of color for cld letters.}

\item{cld_y_adj, cld_y_mult}{(numeric) y position correction (addition and multiplication) factors for cld letters.}

\item{ci_x_adj}{(numeric)  x position correction factor for mean confidence interval.}

\item{points_x_adj}{(numeric) x position correction factor for jittered points.}

\item{...}{arguments to \code{sort_fun}.}
}
\value{
A ggplot2 plot object.
}
\description{
Plot a boxplot with additional components
}
\examples{
# Example 1
gg_boxplot_plus(decrease ~ treatment, OrchardSprays)


# Example 2
gg_boxplot_plus(decrease ~ treatment, OrchardSprays, sort_groups = "yes")


# Example 3a

res <- posthoc_anova(weight ~ Diet, data = ChickWeight)
cld_result <- make_cld(res)

gg_boxplot_plus(weight ~ Diet, data = ChickWeight,
                cld = cld_result)

# Example 3b

gg_boxplot_plus(weight ~ Diet, data = ChickWeight,
                cld = cld_result,
                sort_groups = "descending",
                sort_fun = mean)

# Example 4

res2 <- posthoc_anova(decrease ~ treatment, data = OrchardSprays)
cld_result2 <- make_cld(res2)

gg_boxplot_plus(decrease ~ treatment, data = OrchardSprays,
                cld = cld_result2,
                sort_groups = "descending",
                sort_fun = mean)

gg_boxplot_plus(decrease ~ treatment, data = OrchardSprays,
                cld = cld_result2,
                sort_groups = "ascending",
                sort_fun = IQR)

}
