% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cld.R
\name{make_cld}
\alias{make_cld}
\alias{make_cld.pairwise.htest}
\alias{make_cld.posthocTGH}
\alias{make_cld.posthoc_anova}
\alias{make_cld.PMCMR}
\alias{make_cld.formula}
\alias{make_cld.matrix}
\alias{make_cld.pairwise_pval_df}
\title{Make a compact letter display for pair-wise comparison}
\usage{
make_cld(obj, ..., alpha = 0.05)

\method{make_cld}{pairwise.htest}(obj, ..., alpha = 0.05)

\method{make_cld}{posthocTGH}(obj, ..., alpha = obj$intermediate$alpha)

\method{make_cld}{posthoc_anova}(obj, ..., alpha = obj$intermediate$alpha)

\method{make_cld}{PMCMR}(obj, ..., alpha = 0.05)

\method{make_cld}{formula}(obj, ..., data = NULL, alpha = 0.05)

\method{make_cld}{matrix}(obj, ..., alpha = 0.05)

\method{make_cld}{pairwise_pval_df}(obj, ..., alpha = 0.05)
}
\arguments{
\item{obj}{Object with pair-wise comparisons (e.g., post-hoc test results).
Currently supported objects: \itemize{
\item \emph{posthocTGH} from package \pkg{userfriendlyscience};
\item \emph{PMCMR} from package \pkg{PMCMR}.
}}

\item{...}{Further arguments to methods.}

\item{alpha}{(numeric from 0 to 1) Significance level.}
}
\value{
(A data frame with) compact letter display.
}
\description{
Make a compact letter display for results of pair-wise comparisons (e.g., ANOVA post-hoc tests, Kruskal-Wallis post-hoc tests and other).
}
\details{
This function is based on function \code{\link[rcompanion]{cldList}()} from package \pkg{rcompanion}.
}
\examples{
# Example 1: class `pairwise.htest`

obj1 <- pairwise.wilcox.test(chickwts$weight, chickwts$feed, exact = FALSE)
make_cld(obj1)


# Example 2: class `pairwise.htest`

obj2 <- with(OrchardSprays, pairwise.t.test(decrease, treatment))
make_cld(obj2)


# Example 3: class `PMCMR`

\donttest{
obj3 <- PMCMR::posthoc.kruskal.conover.test(count ~ spray,
                                            data = InsectSprays)
make_cld(obj3)
}


# Example 4: class `posthocTGH`

obj4 <- posthoc_anova_tukey(weight ~ Diet, data = ChickWeight)
make_cld(obj4)


# Example 5: class `posthoc_anova`

obj5 <- posthoc_anova_games_howell(weight ~ Diet, data = ChickWeight)
make_cld(obj5)


# Example 6: class `formula`

DataFrame <- data.table::fread(
    'Comparison     p.value p.adjust
    "EE - GB = 0"        1 1.000000
    "EE - CY = 0" 0.001093 0.003279
    "GB - CY = 0" 0.005477 0.008216')

make_cld(p.adjust ~ Comparison, data = DataFrame)


# Example 7: class `matrix`
# (for symetric matrices of p values)

# Create matrix
m <- c(1.00, 0.22, 0.05, 0.00,
       0.22, 1.00, 0.17, 0.01,
       0.05, 0.17, 1.00, 0.22,
       0.00, 0.01, 0.22, 1.00)
obj7 <- matrix(m, nrow = 4)
rownames(obj7) <- colnames(obj7) <- c("P", "O", "I", "U")
obj7

# Make cld
make_cld(obj7)
}
