% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{plot.qqdata}
\alias{plot.qqdata}
\title{Plot qqdata object as qqplot}
\usage{
\method{plot}{qqdata}(x, ..., use_colors = FALSE, scales = "free")
}
\arguments{
\item{x}{qqdata object}

\item{...}{other parameters}

\item{use_colors}{(logical) TRUE|FALSE use colors for multiple groups}

\item{scales}{(string) "fixed"|"free"|"free_x"|"free_y"}
}
\description{
Plot qqdata object as qqplot
}
\examples{
library(BioStat)
data(chickwts, package = "datasets")

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Input as formula + data:

QQ_groups <- qq_data(weight ~ feed, data = chickwts, method = "normal")
plot(QQ_groups)

QQ_groups <- qq_data(weight ~ feed, data = chickwts, method = "any")
plot(QQ_groups)

plot(QQ_groups, scales = "fixed")
plot(QQ_groups, use_colors = TRUE)


QQ_single <- qq_data( ~ weight, data = chickwts)
plot(QQ_single)

class(QQ_single)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

data(CO2, package = "datasets")

 QQ_CO2 <- qq_data(uptake ~ Type + Treatment, data = CO2)
 plot(QQ_CO2)

 QQ_CO2_B <- qq_data(uptake ~ Type + Treatment, data = CO2, line = "robust")
 plot(QQ_CO2_B)

}
