% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_data.R
\name{plot.qqdata}
\alias{plot.qqdata}
\title{A method to plot a `qqdata` object as a qq-plot}
\usage{
\method{plot}{qqdata}(x, ..., use_colors = FALSE, scales = "free")
}
\arguments{
\item{x}{A \code{qqdata} object}

\item{...}{other parameters}

\item{use_colors}{(logical) use colors for multiple groups}

\item{scales}{("free"|"free_x"|"free_y"|"fixed")
a parmeter to be passed to
 \code{\link[ggplot2]{facet_wrap}}.}
}
\description{
A method to plot a `qqdata` object as a qq-plot
}
\examples{
library(BioStat)
data(chickwts, package = "datasets")

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as formula + data:

QQ_groups <- qq_data(weight ~ feed, data = chickwts, method = "normal")
plot(QQ_groups)

QQ_groups <- qq_data(weight ~ feed, data = chickwts, method = "any")
plot(QQ_groups)

plot(QQ_groups, scales = "fixed")
plot(QQ_groups, use_colors = TRUE)


QQ_single <- qq_data( ~ weight, data = chickwts)
plot(QQ_single)

class(QQ_single)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

data(CO2, package = "datasets")

 QQ_CO2 <- qq_data(uptake ~ Type + Treatment, data = CO2)
 plot(QQ_CO2)

 QQ_CO2_B <- qq_data(uptake ~ Type + Treatment, data = CO2, line = "robust")
 plot(QQ_CO2_B)

}
