% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfc_posthoc_anova.R
\name{posthoc_anova}
\alias{posthoc_anova}
\alias{posthoc_anova.formula}
\alias{posthoc_anova.default}
\alias{print.posthoc_anova}
\title{Games-Howell and Tukey post-hoc tests}
\usage{
posthoc_anova(y, x = NULL, method = c("Games-Howell", "Tukey"),
  conf_level = 0.95, digits = 2, p.adjust = "holm",
  format_pvalue = TRUE, data = NULL, ...)

\method{posthoc_anova}{formula}(y, x = NULL, method = c("Games-Howell",
  "Tukey"), conf_level = 0.95, digits = 2, p.adjust = "holm",
  format_pvalue = TRUE, data = NULL, ..., sep = " | ")

\method{posthoc_anova}{default}(y, x = NULL, method = c("Games-Howell",
  "Tukey"), conf_level = 0.95, digits = 2, p.adjust = "holm",
  format_pvalue = TRUE, data = NULL, ...)

\method{print}{posthoc_anova}(x, digits = x$input$digits, ...)
}
\arguments{
\item{y}{(numeric|formula) Either a numeric vector or formula.}

\item{x}{(factor) a vector that either is a factor or can be converted into
one. If \code{y} is a formula, \code{x} is ignored.}

\item{method}{(\code{"games-howell"}|\code{"tukey"}) Name of post-hoc
tests to conduct. Valid values are "tukey" and "games-howell".}

\item{conf_level}{(number) Confidence level of the confidence intervals. Number from 0 to 1.}

\item{digits}{The number of digits to show in the output.}

\item{p.adjust}{Any valid \code{\link{p.adjust}} method.}

\item{format_pvalue}{Whether to format the p values according to APA
standards (i.e. replace all values lower than .001 with '<.001'). This only
applies to the printing of the object, not to the way the p values are
stored in the object.}

\item{data}{(factor) a data frame with data to be used in combination with formula.}

\item{...}{Further arguments to methods.}

\item{sep}{(character) A string with a symbol to separate group names if several grouping variables are used.}

\item{...}{(further arguments to methods)}
}
\value{
A list of three elements:
\item{input}{List with input arguments}
\item{intermediate}{List of intermediate objects.}
\item{output}{List with
two objects 'tukey' and 'games.howell', containing the outcomes for the
respective post-hoc tests.}
}
\description{
Either Games-Howell test or Tukey honestly significant difference (HSD) post-hoc
tests for one-way analysis of variance (ANOVA). The main part of the
function code and descriptions were
imported from \code{posthocTGH()} in package \pkg{userfriendlyscience}.
}
\note{
This function is based on a file that was once hosted at
\url{http://www.psych.yorku.ca/cribbie/6130/games_howell.R}, but has been removed
since. It was then adjusted for implementation in the
\pkg{userfriendlyscience} package. Jeffrey Baggett needed the
confidence intervals, and so emailed them, after which his updated functions
was used. In the meantime, it appears Aaron Schlegel
(\url{https://rpubs.com/aaronsc32}) independently developed a version with
confidence intervals and posted it on RPubs at
\url{https://rpubs.com/aaronsc32/games-howell-test}.
}
\examples{
library(BioStat)

# Compute post-hoc statistics using the Games-Howell method
posthoc_anova(weight ~ Diet, data = ChickWeight)
posthoc_anova(weight ~ Diet, data = ChickWeight, method = "Games-Howell")


# Compute post-hoc statistics using the Tukey method
posthoc_anova(weight ~ Diet, data = ChickWeight, method = "tukey")


}
\seealso{
\itemize{
\item{\url{http://www.gcf.dkf.unibe.ch/BCB/files/BCB_10Jan12_Alexander.pdf}}
\item{\url{https://rpubs.com/aaronsc32/games-howell-test}}
\item{\url{http://aoki2.si.gunma-u.ac.jp/R/src/tukey.R}}
\item{\url{https://stats.stackexchange.com/questions/83941/games-howell-post-hoc-test-in-r}}

}
}
\author{
Gjalt-Jorn Peters (Open University of the Netherlands) & Jeff Bagget
(University of Wisconsin - La Crosse)

Maintainer: Gjalt-Jorn Peters <gjalt-jorn@userfriendlyscience.com>
}
\keyword{utilities}
