% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_data.R
\name{qq_data}
\alias{qq_data}
\alias{coef.qqdata}
\title{Calculate data for a qq-plot}
\usage{
qq_data(x, data = NULL, distribution = "norm", ..., envelope = 0.95,
  line = c("quartiles", "robust", "int=0,slope=1", "0,1", "none"),
  labels = NULL, groups = NULL, method = c("mle-normal", "trimmed-normal",
  "moment-normal", "any"), sep = " | ")

\method{coef}{qqdata}(object, ...)
}
\arguments{
\item{x}{Either a formula, a numeric vector or a name of a vector
in \code{data}.
If \code{x} is a formula (e.g. \code{variable ~ factor}), left-hand
side provides variable to be summarized. Right-hand side and condition
describe subsets. If the left-hand side is empty, right-hand side and
condition are shifted over as a convenience.}

\item{data}{A data frame that contains the variables mentioned in \code{x}.}

\item{distribution}{root name of comparison distribution -- e.g., \code{"norm"} for the
    normal distribution; \code{t} for the t-distribution.}

\item{...}{Parameters to be passed to function, selected in \code{distribution}.}

\item{envelope}{(numeric) confidence level for point-wise confidence envelope.}

\item{line}{(string) A parameter, that controls how reference line is drawn.
           Options:\itemize{
           \item{\code{"0,1"} or \code{"int=0,slope=1"} to plot a line
            with intercept = 0 and slope = 1;}
           \item{ \code{"quartiles"} to pass a line through the quartile-pairs;}
           \item{ \code{"robust"} for a robust-regression line;
            the latter uses the \code{rlm} function in the \pkg{MASS} package};
           \item{option \code{"none"} is not implemented yet.}
}}

\item{labels}{vector of text strings to be used to identify points, defaults to
    \code{names(x)} or observation numbers if \code{names(x)} is \code{NULL}.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}

\item{method}{(string: \code{"trimmed-normal"}, \code{"normal"},
             \code{"any"}). A method, that controls how parameters for
             \code{distribution} are computed.
     Options:\itemize{

           \item{\code{"mle-normal"} (default) all data values are used to
           calculate parameters of theoretical normal distribution using
           method of maximum likelihood;}

           \item{\code{"trimmed-normal"} 10\% of the most extreme
           data values are trimmed before parameters of theoretical normal
           distribution are calculated using method of moments;}

           \item{\code{"moment-normal"} all data values are used to calculate
           parameters of theoretical normal distribution using method of moments;}

           \item{\code{"any"} parameters ate provided manually by user.}
     }
          Options \code{"mle-normal"}, \code{"trimmed-normal"} and
          \code{"moment-normal"} are applicable only if
          \code{distribution = "norm"}.
          Otherwise \code{"any"} is selected automatically.}

\item{sep}{(not used yet).}

\item{object}{A \code{qqdata} object.}
}
\value{
An object, which inherits from classes \code{qqdata} and
        \code{data.frame}. The object contains information, needed
        to plot a qqplot with reference line and its confidence intervals.
        These variables are contained: \itemize{
        \item\strong{x} – x axis values;
        \item\strong{y} – y axis values for points of qq plot;
        \item\strong{labels} – labels for each point;
        \item\strong{ref_y} – y axis values for reference line
        \item\strong{ref_ci_upper} and \strong{ref_ci_lower}
          – y axis values for upper and lower pointwise
           confidence interval of a reference line.
        }
}
\description{
Calculate data for a qq-plot
}
\examples{

library(BioStat)
data(chickwts, package = "datasets")

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as formula + data:

QQdata <- qq_data(~weight, data = chickwts)
head(QQdata)
coef(QQdata)

# Column ".group" is added if applied by group:

QQdata <- qq_data(weight ~ feed, data = chickwts)
head(QQdata)
coef(QQdata)

qq_plot(weight ~ feed, data = chickwts)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as variable name + data:

QQdata <- qq_data("weight", data = chickwts)
head(QQdata)
coef(QQdata)

QQdata <- qq_data("weight",groups = "feed", data = chickwts)
head(QQdata)
coef(QQdata)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as vector

QQdata <- qq_data(chickwts$weight)
head(QQdata)
coef(QQdata)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as vector, several groups.
# Column ".group" is added

QQdata <- qq_data(chickwts$weight, groups = chickwts$feed)
head(QQdata)
coef(QQdata)



}
\seealso{
\code{\link[car]{qqPlot}} in \pkg{car} package,
 \code{\link[stats]{qqplot}} in \pkg{stats} package.
}
