% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_data.R
\name{qq_data}
\alias{qq_data}
\alias{coef.qqdata}
\title{[!]Compute data for a qq-plot}
\usage{
qq_data(x, data = NULL, distribution = "norm", ..., line = c("quartiles",
  "robust", "int=0,slope=1", "0,1", "none"), envelope = 0.95,
  method = c("mle-normal", "trimmed-normal", "moment-normal", "any"),
  labels = NULL, groups = NULL, sep = " | ")

\method{coef}{qqdata}(object, ...)
}
\arguments{
\item{x}{Either a formula, a numeric vector or a name of a column
            in \code{data}. \itemize{
    \item If \code{x} is a formula (e.g. \code{variable ~ factor}), left-hand
    side provides the name of a variable to be tested. In the right-hand side
    there are names of factor variables to be used to create subsets.
    If the left-hand side is empty ((e.g. \code{~ factor}), right-hand
                                    side is treated as variable name to test.
}}

\item{data}{(data frame|\code{NULL}) Either a data frame that contains the
variables mentioned in \code{x} or \code{NULL} (if the variables
are in the function's environment).}

\item{distribution}{root name of comparison distribution -- e.g., \code{"norm"} for the
    normal distribution; \code{t} for the t-distribution.}

\item{...}{Parameters to be passed to function, selected in \code{distribution}.}

\item{line}{(string) A parameter, that controls how reference line is drawn.
           Options:\itemize{
    \item{\code{"0,1"} or \code{"int=0,slope=1"} to plot a line
            with intercept = 0 and slope = 1;}
    \item{ \code{"quartiles"} to pass a line through the quartile-pairs;}
    \item{ \code{"robust"} for a robust-regression line;
            the latter uses the \code{rlm} function from the \pkg{MASS} package};
    \item{option \code{"none"} is not implemented yet.}
}}

\item{envelope}{(numeric | \code{FALSE}) confidence level for point-wise
confidence envelope.}

\item{method}{(string: \code{"trimmed-normal"}, \code{"normal"},
             \code{"any"}). A method, that controls how estimates of
             parameters for \code{distribution} are computed.
     Options:
     \itemize{

           \item{\code{"mle-normal"} (default) all data values are used to
           estimate parameters of theoretical normal distribution using
           method of maximum likelihood;}

           \item{\code{"trimmed-normal"} 10\% of the most extreme
           data values are trimmed before parameters of theoretical normal
           distribution are estimated using method of moments;}

           \item{\code{"moment-normal"} all data values are used to estimate
           parameters of theoretical normal distribution using method of moments;}

           \item{\code{"any"} either parameters are provided manually by user
           or default parameters are used (if any).}
     }
          Options \code{"mle-normal"}, \code{"trimmed-normal"} and
          \code{"moment-normal"} are applicable only if
          \code{distribution = "norm"}.
          Otherwise \code{"any"} is selected automatically.}

\item{labels}{vector of text strings to be used to identify points, defaults to
    \code{names(x)} or observation numbers if \code{names(x)} is \code{NULL}.}

\item{groups}{(\code{NULL}|factor|character) An alternative way to provide
groups. If \code{x} is a numeric vector, \code{groups} must be
a factor (or \code{NULL}). If \code{x} is a sting,
\code{groups} must also be a string (or \code{NULL}).}

\item{sep}{(not used yet).}

\item{object}{A \code{qqdata} object.}
}
\value{
An object, which inherits from classes \code{qqdata} and
        \code{data.frame}. The object contains information, needed
        to plot a qqplot with reference line and its confidence intervals.
        These variables are contained: \itemize{
        \item\strong{x} – x axis values;
        \item\strong{y} – y axis values for points of qq plot;
        \item\strong{labels} – labels for each point;
        \item\strong{ref_y} – y axis values for reference line
        \item\strong{ref_ci_upper} and \strong{ref_ci_lower}
          – y axis values for upper and lower pointwise
           confidence interval of a reference line.
        }
}
\description{
Compute data necessary to plot a quntile comparison plot (qq-plot).
}
\details{
Function \code{qq_data} is inspired by \code{qqPlot()} in package \pkg{car}
(writen by J. Fox).
}
\examples{

library(BioStat)
data(chickwts, package = "datasets")

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as formula + data:

my_qq_df <- qq_data(~weight, data = chickwts)
head(my_qq_df)
coef(my_qq_df)

# Column ".group" is added if applied by group:

my_qq_df <- qq_data(weight ~ feed, data = chickwts)
head(my_qq_df)
coef(my_qq_df)

qq_plot(weight ~ feed, data = chickwts)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as variable name + data:

my_qq_df <- qq_data("weight", data = chickwts)
head(my_qq_df)
coef(my_qq_df)

my_qq_df <- qq_data("weight",groups = "feed", data = chickwts)
head(my_qq_df)
coef(my_qq_df)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as vector

my_qq_df <- qq_data(chickwts$weight)
head(my_qq_df)
coef(my_qq_df)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Input as vector, several groups.
# Column ".group" is added

my_qq_df <- qq_data(chickwts$weight, groups = chickwts$feed)
head(my_qq_df)
coef(my_qq_df)



}
\seealso{
\code{\link[car]{qqPlot}} in \pkg{car} package,
 \code{\link[stats]{qqplot}} in \pkg{stats} package.
}
