% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_line_coeffs.R
\name{qq_line_coeffs}
\alias{qq_line_coeffs}
\title{Get slope and intercept of qqline}
\usage{
qq_line_coeffs(y, datax = FALSE, distribution = qnorm, probs = c(0.25,
  0.75), qtype = 7)
}
\arguments{
\item{y}{The second or only data sample.}

\item{datax}{logical. Should data values be on the x-axis?}

\item{distribution}{quantile function for reference theoretical distribution.}

\item{probs}{numeric vector of length two, representing probabilities.
    Corresponding quantile pairs define the line drawn.}

\item{qtype}{the \code{type} of quantile computation used in \code{\link{quantile}}.}
}
\value{
list with $slope and $intercept for qqline
}
\description{
Get slope and intercept of qqline
}
