% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{A qq-plot for multiple groups (with ggplot2)}
\usage{
qq_plot(x, data = NULL, distribution = "norm", ..., envelope = 0.95,
  line = c("quartiles", "robust", "int=0,slope=1"), labels = NULL,
  groups = NULL, method = if (distribution == "norm") {     "normal" } else
  {     "any" }, use_colors = FALSE, scales = "free")
}
\arguments{
\item{x}{Either a formula or a numeric vector or a name of a vector
in \code{data}.\cr
If \code{x} is a formula (e.g. \code{variable ~ factor}), left
side provides variable to be summarized. Right side and condition
describe subsets. If the left side is empty, right side and
condition are shifted over as a convenience.}

\item{data}{A data frame that contains the variables mentioned in \code{x}.}

\item{distribution}{root name of comparison distribution -- e.g., \code{"norm"} for the
    normal distribution; \code{t} for the t-distribution.}

\item{...}{Parameters to be passed to function, selected in \code{distribution}}

\item{envelope}{confidence level for point-wise confidence envelope, or 
    \code{FALSE} for no envelope.}

\item{line}{\code{"quartiles"} to pass a line through the quartile-pairs, or
    \code{"robust"} for a robust-regression line; the latter uses the \code{rlm}
    function in the \code{MASS} package. Specifying \code{line = "none"} suppresses the line.}

\item{labels}{vector of text strings to be used to identify points, defaults to
    \code{names(x)} or observation numbers if \code{names(x)} is \code{NULL}.}

\item{use_colors}{(logical) use colors for multiple groups}

\item{scales}{("free"|"free_x"|"free_y"|"fixed")
a parmeter to be passed to
 \code{\link[ggplot2]{facet_wrap}}.}
}
\description{
A qq-plot for multiple groups (with ggplot2)
}
\examples{
library(BioStat)
data(iris)

# Formula (several groups):
qq_plot(Sepal.Length ~ Species, data = iris)

# Formula (several groups in colors):
qq_plot(Sepal.Length ~ Species, data = iris, use_colors = TRUE)

# Vectors (several groups):
qq_plot(iris$Sepal.Length, groups = iris$Species)

# For one group:
qq_plot("Sepal.Length", data = iris)

qq_plot(~Sepal.Length, data = iris)

qq_plot(iris$Sepal.Length)

}
