% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{test_normality}
\alias{test_normality}
\alias{print.test_normality}
\title{Normality tests with formula interface}
\usage{
test_normality(x, data = NULL, test = "Shapiro-Wilk", ..., groups = NULL)

\method{print}{test_normality}(x, ..., digits = 3)
}
\arguments{
\item{x}{Either a formula or a numeric vector or a name of a vector
in \code{data}.\cr
If \code{x} is a formula (e.g. \code{variable ~ factor}), left
side provides variable to be summarized. Right side and condition
describe subsets. If the left side is empty, right side and
condition are shifted over as a convenience.}

\item{data}{A data frame that contains the variables mentioned in \code{x}.}

\item{test}{(string | function) Either a string  (case insensitive, may be
            unambiguously abbreviated) with a name of nomality test or a
            function, which carries out the test.\cr
Possible names of tests:\itemize{
\item{"SW", "Shapiro-Wilk" — for Shapiro-Wilk test;}
\item{"Lilliefors" — for Kolmogorov-Smirnov test wtih Lilliefor's correction;}
\item{"AD", "Anderson-Darling" — for Anderson-Darling test;}
\item{"CVM", "CM", "Cramer-von Mises" — for Cramer-von Mises test;}
\item{"SF", "Shapiro-Francia" — for Shapiro-Francia test;}
\item{"Chi-squared","Pearsons" — for Pearson's chi-squared test of normality.}
}}

\item{...}{Parameters to be passed to the main function
(defined/selected in \code{fun}) that carries out a normality test.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}

\item{digits}{A number of significant digits to be displayed. Default is 3.}
}
\value{
A data frame with normality test
         results for each group.
}
\description{
Perform Shapiro-Wilk (default),
Lilliefors (Kolmogorov-Smirnov), Anderson_darling and other tests of normality.
}
\examples{
library(BioStat)
library(pander)
data(CO2)

rez <- test_normality(uptake ~ Type + Treatment,
                      data = CO2)
rez

rez2 <- test_normality(uptake ~ Type + Treatment,
                       data = CO2,
                       test = "chi-sq")
rez2

# Prettify the output
pander(rez)

rez \%>\%
    prettify_p_column() \%>\%
    pander()

format_numbers(rez, 3)

rez \%>\% format_numbers(3)

rez \%>\%
   format_numbers(3) \%>\%
   pander()
}
\seealso{
Package \pkg{nortest}\cr
}
