% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{test_normality}
\alias{test_normality}
\alias{print.normality_test}
\title{Normality tests with formula interface}
\usage{
test_normality(x, test = "Shapiro-Wilk", ..., data = NULL, groups = NULL)

\method{print}{normality_test}(obj, ..., digits = 3)
}
\arguments{
\item{x}{a numeric vector of data values. Missing values are allowed,
    but the number of non-missing values must be between 3 and 5000.}

\item{test}{(string | function) Either one of the following string (case insensitive, may be unambiguously abbreviated) which describes a desired test of normality:\cr
"SW", "Shapiro-Wilk",\cr
"Lilliefors",\cr
"AD", "Anderson-Darling",\cr
"CVM", "CM", "Cramer-von Mises",\cr
"SF", "Shapiro-Francia",\cr
"Chi-square","Pearson"  \cr

or a function, which carries out the test.}

\item{...}{}

\item{data}{a data frame.
Note that the default is \code{data=parent.frame()}.  This makes it convenient to
use this function interactively by treating the working envionment as if it were 
a data frame.  But this may not be appropriate for programming uses.  
When programming, it is best to use an explicit \code{data} argument
-- ideally supplying a data frame that contains the variables mentioned
in \code{formula}.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}

\item{obj}{An object of class \code{normality_test}.}

\item{digits}{(numeric) number of significant digits to display.
Default is 3.}
}
\value{
A data frame with results for each group

                 /NOT DESCRIBED YET/
}
\description{
Performs the Shapiro-Wilk (default), Lilliefors (Kolmogorov-Smirnov) and other test of normality.

Print `normality_test` object
}
\examples{
library(BioStat)
data(CO2)
test_normality(uptake ~ Type + Treatment, data = CO2)

test_normality(uptake ~ Type + Treatment, "chi-sq", data = CO2)
}
