% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\alias{print.test_normality}
\alias{pander.test_normality}
\title{Normality tests with formula interface}
\usage{
test_normality(x, data = NULL, test = "Shapiro-Wilk", ..., groups = NULL,
  sep = " | ")

\method{print}{test_normality}(x, ..., rm_col_method = TRUE)

\method{pander}{test_normality}(x, ..., caption = NA, rm_col_method = TRUE)
}
\arguments{
\item{x}{Either a formula, a numeric vector or a name of a vector
in \code{data}.
If \code{x} is a formula (e.g. \code{variable ~ factor}), left-hand
side provides variable to be summarized. Right-hand side and condition
describe subsets. If the left-hand side is empty, right-hand side and
condition are shifted over as a convenience.}

\item{data}{A data frame that contains the variables mentioned in \code{x}.}

\item{test}{(string | function) Either a string  (case insensitive, may be
            unambiguously abbreviated) with a name of nomality test or a
            function, which carries out the test.\cr
Possible names of tests:\itemize{
\item{"SW", "Shapiro-Wilk" — for Shapiro-Wilk test;}
\item{"Lilliefors" — for Kolmogorov-Smirnov test wtih Lilliefor's correction;}
\item{"AD", "Anderson-Darling" — for Anderson-Darling test;}
\item{"CVM", "CM", "Cramer-von Mises" — for Cramer-von Mises test;}
\item{"SF", "Shapiro-Francia" — for Shapiro-Francia test;}
\item{"Chi-squared","Pearsons" — for Pearson's chi-squared test of normality.}
}}

\item{...}{Parameters to be passed to the main function
(defined/selected in \code{fun}) that carries out a normality test.}

\item{groups}{grouping variable that will be folded into the formula (if there is room for it).
This offers some additional flexibility in how formulas can be specified.}

\item{sep}{(not used yet)}

\item{rm_col_method}{(logical) If \code{TRUE} column "method" is not printed.}

\item{caption}{(string|\code{NULL}|\code{NA}) A caption for the table with results of a normality test. If \code{NA} — a default caption is printed (default). If \code{NULL} – no caption is printed.}
}
\value{
A data frame with normality test
         results for each group.
}
\description{
Perform Shapiro-Wilk (default),
Lilliefors (Kolmogorov-Smirnov), Anderson_darling and other tests of normality.
}
\details{
By default, methods \code{print.test_normality} and
         \code{pander.test_normality} do not print column called "method".
}
\examples{
library(BioStat)
library(pander)
data(CO2)

rez <- test_normality(uptake ~ Type + Treatment,
                      data = CO2)
rez

rez2 <- test_normality(uptake ~ Type + Treatment,
                       data = CO2,
                       test = "chi-sq")
rez2

# Round to 3 decimal digits (all columns):
format_numbers(rez, 3)

# Round to 3 significant digits (all columns):
format_numbers(rez, 3, format = "g")

# Prettify the column with p-values
prettify_p_column(rez)

# Print as a 'pandoc' table
pander(rez)

# Other output possibilities
rez \%>\% format_numbers(3)

rez \%>\%
   format_numbers(3) \%>\%
   pander()

rez \%>\%
    format_numbers(3, "g") \%>\%
    prettify_p_column() \%>\%
    pander()
}
\seealso{
Package \pkg{nortest}\cr
}
